moved PasswordValidationUtil logic into ValidationUtil
