+++ data/users.csv
@@ -1,6 +1,7 @@
 nric,passwordHash,name,age,maritalStatus,role
 S8877665J,8aM5flUu7eF5oQzLPlna0Q==$10000$x1JQNFfaAYJdWahLWTBXHF+ShnJz9VwsDo1MgbYtdLA=,Alan,30,Married,HDB_OFFICER
 S4321098C,xDByXrQ26+5F7PdCiqGA9Q==$10000$yIWShhHWujT/EjHpkNrTizMKf6BX+KaqY3Zxm2XD/PY=,Alex,21,Married,APPLICANT
+S2121212A,eG8Zo+/cDHbw/VAADyr3Lw==$10000$LByfVCL03x/m68JLdM+IGWN5CN6rsi+rBb+Y4Aj2TIw=,Bob,22,Single,APPLICANT
 T1234567J,8aM5flUu7eF5oQzLPlna0Q==$10000$x1JQNFfaAYJdWahLWTBXHF+ShnJz9VwsDo1MgbYtdLA=,David,29,Married,HDB_OFFICER
 T7654321B,xDByXrQ26+5F7PdCiqGA9Q==$10000$yIWShhHWujT/EjHpkNrTizMKf6BX+KaqY3Zxm2XD/PY=,Sarah,40,Married,APPLICANT
 S5678901G,ekmaze7bxEvLLbQNIn3Fcg==$10000$43PRiB6uTLjxC+c8LdvvfHppqRcCiBgokZH6zy9gI+4=,Jessica,26,Married,HDB_MANAGER
